Yarn gen failed hard when importing this function, so moving it to a separate file is probably better
