+++ src/layout/Likert/Generator/makeLikertChildId.ts
@@ -1,0 +1,6 @@
+export function makeLikertChildId(parentId: string, rowIndex: number | undefined) {
+  if (rowIndex === undefined) {
+    return `${parentId}-item`;
+  }
+  return `${parentId}-item-${rowIndex}`;
+}
